pyproject.toml
[project]
name = "contentforge-backend"
version = "0.1.0"
description = "Fast and efficient backend for ContentForge, a content generation platform."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.83.0",
    "fastapi>=0.129.2",
    "google-genai>=1.65.0",
    "newspaper4k>=0.9.4.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.41.0",
]
